Renamed request package to msg
to avoid clashes with other things named `request`
