+++ msg/doc.go
@@ -1,2 +1,2 @@
 // Generic HTTP responses that aren't specific to any particular type of request
-package response
+package msg
