+++ msg/message.go
@@ -1,4 +1,4 @@
-package response
+package msg
 
 import (
 	"fmt"
